backend/requirements.txt
flask
flask-cors
google-genai
duckduckgo-search
pdfminer.six
requests
